Example file
This CMD file will read the name set to this file and set that as the computer name to be used by the Window update master script. 

How To:-

Create a new text file and name it to match another device connected to your network. Open the file and copy the contents of this file into it.